requirements.txt
Django==1.10
github3.py==1.0.0a4
python-decouple==3.0
